Fix incorrect partial message detection threshold
Prior to this commit, the logic to determine that a Kafka broker
had returned a partial message was incorrectly calculating and
checking for how much data needed to be available to decode a
message. This would result in an "error decoding remaining
message bytes" error if the size of the partial message was
less than 4 bytes of the complete message.

This commit resolves this bug by properly calculating and
checking for how much data needs to be available for decoding
a message.